chore(cc): rename domain files to kebab-case
