+++ projects/ng-payment-card/src/lib/domain/card-types.ts
@@ -1,4 +1,4 @@
-import { CardType } from './card-type';
+import { CardType } from './card-type.enum';
 
 const CARD_TYPES: Map<string, RegExp> = new Map();
 CARD_TYPES.set(CardType.AMERICAN_EXPRESS, new RegExp('^3[47]'));
