+++ projects/ng-payment-card/src/lib/domain/i-card-details.ts
